Moving types for AST into tree.go
